refactor: 테스트에서만 사용되는 구현체의 위치를 이동합니다. AlwaysGenerateBridgeStrategy main -> test
