+++ src/test/java/domain/ladder/strategy/AlwaysGenerateBridgeStrategy.java
